Refactoring!
