+++ src/main/java/com/kamesuta/mc/modspawner/gui/logo.png
