+++ src/main/java/com/kamesuta/mc/modspawner/launch/Loader.java
@@ -11,9 +11,11 @@ import java.net.URLConnection;
 //import org.apache.logging.log4j.LogManager;
 //import org.apache.logging.log4j.Logger;
 import com.kamesuta.mc.modspawner.asm.MdspCorePlugin;
+import com.kamesuta.mc.modspawner.gui.DLGui;
+import com.kamesuta.mc.modspawner.gui.DownloadMonitor;
+import com.kamesuta.mc.modspawner.gui.DownloadMonitor.IDownloadCloser;
+import com.kamesuta.mc.modspawner.gui.DownloadMonitor.IDownloadProgress;
 import com.kamesuta.mc.modspawner.launch.DLCalculate.DLDetails;
-import com.kamesuta.mc.modspawner.launch.DownloadMonitor.IDownloadCloser;
-import com.kamesuta.mc.modspawner.launch.DownloadMonitor.IDownloadProgress;
 
 public class Loader {
 	// private static ByteBuffer downloadBuffer = ByteBuffer.allocateDirect(1 <<
